feat: added previews
